requirements.txt
atproto==0.0.53
peewee~=3.16.2
Flask~=2.3.2
python-dotenv~=1.0.0
openai==1.54.5
launchdarkly-server-sdk==9.8.0